Create Danis&Probs.c
Daanish as always is busy creating and solving his favourite and interesting graph problems. Chef assigns each problem a difficulty — an integer between 1 and 10. For each valid i, there are Ai problems with difficulty i

.

A participant hacked Daanish's account and got access to the problem proposal document. He can delete up to K
problems from the document in order to reduce the difficulty of the contest for himself and his friends. Find the smallest possible value of the difficulty of the most difficult problem which remains after removing K problems.

Explanation

Example case 1: The participant can remove all 100
problems with difficulty 10. Then, there are no problems with difficulties 9 and 10, so the maximum difficulty among remaining problems is 8

.

Example case 2: The participant can remove all problems with difficulties 3
and 10 and any one of the problems with difficulty 1. Then, the only problem left has difficulty 1

.

Example case 3: The participant cannot remove any problem. The document does not contain any problems with difficulties 9
or 10, so the maximum difficulty of a problem is 8.