Create COLGLF2.c
Chef has just started watching Game of Thrones, and he wants to first calculate the exact time (in minutes) that it'll take him to complete the series.

The series has S seasons, and the ith season has Ei episodes, each of which are Li,1,Li,2,…,Li,Ei minutes long. Note that these Li,j include the duration of the beginning intro song in each episode. The streaming service that he uses, allows Chef to skip the intro song. The intro song changes slightly each season, and so he wants to watch the intro song in the first episode of each season, but he'll skip it in all other episodes of that season (yes, we know, a sacrilege!). You know that the intro song lasts for Qi minutes in the ith season.

Find the total time in minutes, that he has to watch.

Explanation:
1 in the beginning denotes there is only 1 test case.

Testcase 1:

There are 2 seasons. The intro song in each of the first season episodes lasts for 1 minute, and the intro song in the second season episodes lasts for 2 minutes each.

For the first season, since there is only 1 episode, Chef will be watching it completely - 2 minutes.

For the second season, Chef will be watching the first episode completely (3 minutes) and will skip the intro song of the second episode (4−2=2 minutes).

So, the total time spent is 2+(3+(4−2))=7 minutes.