Create COUPON2.c
Chef wants to order food from a food delivery app. He wishes to order once today, and buy three items costing A1,A2 and A3 rupees, respectively. He'll also order once tomorrow, when he'll buy three items costing B1,B2 and B3 rupees, respectively. There is an additional delivery charge of rupees D for each order.

He notices that there is a coupon on sale, which costs rupees C. If he buys that coupon, the delivery charges on any day, on an order of rupees 150 or more shall be waived (that is, the D rupees will not be added, if the sum of the costs of the items is ≥150).

Note that Chef is ordering the three items together on each day, so the delivery charge is charged only once each day. Also, note that it's only needed to buy the coupon once to avail the delivery fee waiver on both days.

Should Chef buy the coupon? Note that Chef shall buy the coupon only if it costs him strictly less than what it costs him without the coupon, in total.

Explanation:
Testcase 1: On the first day, Chef plans on ordering three food items, costing a total of rupees (100+50+10)=160. On the second day, Chef plans on ordering three food items, costing a total of rupees (80+80+80)=240.

If Chef buys the coupon he shall save the delivery charge on both days, but the coupon shall cost him rupees 100 extra. In total, it shall cost him (160+240+100)=500.

Without the coupon, Chef will have to pay the delivery charges for both days, which shall cost him a total of rupees (160+240+90+90)=580.

Thus, it's better for Chef to buy the coupon.

Testcase 2: On the first day, Chef plans on ordering three food items, costing a total of rupees (100+100+100)=300. On the second day, Chef plans on ordering three food items, costing a total of rupees (10+20+30)=60.

If Chef buys the coupon he shall save the delivery charge on only the first day, but the coupon shall cost him rupees 30 extra. In total, it shall cost him (300+60+30 (coupon cost) +30 (delivery cost for second day))=420.

Without the coupon, Chef will have to pay the delivery charges for both days, which shall cost him a total of rupees (300+60+30 (delivery cost for first day)+30 (delivery cost for second day))=420.

Since it shall cost him the same amount in both the cases, Chef shall not buy the coupon.